Added routine to treat potential symmetry breaking
If the supercell has broken symmetry, the equivalent pckpts are no
longer equivalent. They should be sampled in the SC and have average
taken in the end. The remaining symmetry operations in the supercell
can be used to reduce the set of extra kpoints used in the sampling.

We construct a reduced set of supercell kpoints while keep a mapping
between it and the origin set.